Choose bundle with highest version number when no default bundle is set
